requirements.txt
pymysql
termcolor
python-binance
